requirements.txt
python-dotenv
semantic-kernel[mcp]
mcp
